requirements_segmentation.txt
dotmap==1.3.30
matplotlib==3.3.3
numpy==1.23.4
pandas==1.5.1
Pillow==9.3.0
pycocotools==2.0.5
pydantic==1.10.2
python-decouple==3.6
python-dotenv==0.21.0
pytorch_lightning==1.7.7
PyYAML==6.0
rasterio==1.3.2
scikit_image==0.19.3
seaborn==0.11.2
Shapely==1.8.5.post1
skimage==0.0
torchgeo==0.3.1
torchmetrics==0.9.3
tqdm==4.64.1
